examples/action_dann_lightn/requirements.txt
# https://github.com/kornia/kornia/blob/master/requirements.txt
# Versions used in development

numpy==1.19.2
pandas==1.2.2
python==3.8

# core dependencies
pytorch==1.7.1
pytorch-lightning==1.2.1
scikit-learn==0.24.1

# utils
torchvision==0.8.2
yacs==0.1.6
